Introduce a separate folder for 'hazards'
