+++ src/game_logic/entity_factory.cpp
@@ -44,12 +44,12 @@
 #include "game_logic/enemies/security_camera.hpp"
 #include "game_logic/enemies/simple_walker.hpp"
 #include "game_logic/enemies/slime_blob.hpp"
-#include "game_logic/enemies/slime_pipe.hpp"
-#include "game_logic/enemies/smash_hammer.hpp"
 #include "game_logic/enemies/snake.hpp"
 #include "game_logic/enemies/spider.hpp"
 #include "game_logic/enemies/spike_ball.hpp"
 #include "game_logic/enemies/watch_bot.hpp"
+#include "game_logic/hazards/slime_pipe.hpp"
+#include "game_logic/hazards/smash_hammer.hpp"
 #include "game_logic/interactive/blowing_fan.hpp"
 #include "game_logic/interactive/elevator.hpp"
 #include "game_logic/interactive/enemy_radar.hpp"
