+++ src/game_logic/hazards/slime_pipe.hpp
