+++ src/game_logic/hazards/smash_hammer.cpp
